+++ src/control/high_level_control.c
@@ -37,7 +37,7 @@
 #include <float.h>
 
 #include "control/high_level_control.h"
-#include "control/motors_PID.h"
+#include "control/motors.h"
 #include "communication/serial.h"
 #include "communication/parsing_messages.h"
 #include "system/user.h"
