+++ src/system/interrupts.c
@@ -37,7 +37,7 @@
 #include "system/system.h"
 #include "communication/serial.h"
 #include "communication/parsing_messages.h"
-#include "control/motors_PID.h"
+#include "control/motors.h"
 #include "control/high_level_control.h"
 
 /******************************************************************************/
